moved examples
