+++ examples/mint_simple_nft.ts
