+++ examples/pool_registration.ts
