+++ examples/typed_data.ts
